Refactor timer script
Timer html is now completely drawn in the `CountDownTimer.drawTimer()`
method. `CountDownTimer.showRemainingTime()` now only updates the element
containing the number that changed.

The countdown itself is now made up of a `<ul>` with each list element
holding a `<div>` for the displayed number and one for it's label.

`<div>`s holding the numbers now have IDs to more easily target.